backend/requirements.txt

Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
django-firebase==0.0.7
django-firebase-orm==0.6.1
djangorestframework==3.15.1

firebase-admin==2.13.0


pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.29.0

sqlparse==0.4.4

uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0

